+++ src/main/java/dev/bungrudi/loanengine/LoanFactory.java
@@ -1,7 +1,8 @@
 package dev.bungrudi.loanengine;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface LoanFactory {
-    Loan createLoan(String loanId, double loanAmount, double interestRate, int numberOfWeeks, LocalDate startDate);
+    Loan createLoan(String loanId, BigDecimal loanAmount, BigDecimal interestRate, int numberOfWeeks, LocalDate startDate);
 }
